refactor: move glob
